Part 1 - Complete
Code for Part 1 complete. Includes the following:
- findAndSetTokenDetails
- removeTokensOfType

Need to perform vigorous testing on it.  Note that identifier algorithm
does not properly parse a list of signal/variable declarations in 1
line. Once complete, we can look at implementing the algorithm to pick
that up. The algorithm.txt file has been updated to do it.

removeTokensOfType has not been tested yet.