Part 2
- Drafted algorithm for Part 2
- Implemented algorithm for Part 2
- Moved stringLower function out of TokenList class so it can be used
elsewhere. May consider placing inside unnamed namespace to keep it
valid inside this compilation unit.
- Modified TokenList::append() function to accommodate appending copied
tokens that already have their previous and next pointers set. This
allows for these pointers to be reset as soon as they are appended to a
new list so that they do not cause conflict with other lists.